+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.4.0 (Unreleased)
+## 1.4.0 (January 14, 2019)
 
 BACKWARDS INCOMPATIBILITIES / NOTES: